Add ability to move O.C. from lect to right
Currently, the mps function leaves the orthogonality center on
the last (leftmost) site in the chain. set_orthogonality adds
capability to move the orthogonality center to any given site
so that expectation values of local operators may be quickly
evaluated in a numerically stable manner. This commit adds
the ability to move the O.C. to any desired location
*given that the O.C. is currently on the leftmost site* and
adds corresponding tests. TODO: add tests for moving O.C. from
right to left.